platform: Add platform default UART control definition
Add platform default UART control definition.
For platforms that wishes to use a UART configuration different from
the default CMSIS control value the default values need to be supplied
in the call to set control.

This makes it possible for a platform to set for example Hardware
Flow Control configuration.
Having support for Hardware Flow Control is useful for running the TF-M
regression tests reliably in a CI system.

Change-Id: I797c5a66a71bc8167bee2f051073d783d13dea63
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>