SPM: Set connection->caller_data to NULL when called through ns_agent_tz
connection->caller_data is intended for use only by the mailbox code.
It only exists when ns_agent_mailbox is ON.
Currently, though, if it exists then it gets populated for every
non-secure request, regardless of whether it arrives through
ns_agent_tz or ns_agent_mailbox.
Fix this by checking the calling partition and setting it to NULL if
not called from ns_agent_mailbox.

Also, in is_tfm_rpc_msg(), check caller_data too and return false
if caller_data is not set.

Without this patch, if TF-M is built with support for both mailbox and
TZ-based NSPEs, the TZ-based one will fail because is_tfm_rpc_msg() will
return true for requests from the TZ-based NSPE.

Signed-off-by: Chris Brand <chris.brand@cypress.com>
Change-Id: I26f6ef485f475e33f7ee1907f436043531a4ad82