Build: Non-FFM manifest attribute needs to be registered
Non-FFM manifest attribute keywords are now used in some manifests.
Those Secure Partitions are TF-M specific and not expected to be
portable. Secure partition developers may take those as references
without being aware of the portability issues.
This patch adds the feature to mandate TF-M-specific Secure Partitions
registering the specific attributes they use. The manifest tool will
report errors if Non-FFM keywords that are not registerd specifically
are detected.
Note that this does not mean using TF-M-specific attributes is
encouraged as those attributes might be specific to that Secure
Partition only.

Signed-off-by: Chendi Sun <chendi.sun@arm.com>
Change-Id: I7c6400eb1fec4239d889634a07da04154e722175