Platform: Fix order of sections in GCC linker script
When section is placed into MEMORY area location counter (.) is updated
with new end address if the address is bigger than previous one.
Also location counter can not go backwards.

Considering following scenario:

Section1 -> FLASH (. is updated)
Section2 -> RAM   (. is updated)
Section3 -> FLASH (. is NOT updated because address of FLASH is
                    smaller than address of RAM, so at this point
                    . points to address in RAM)

In such case location counter can not be used to set absolute address of
Section3 because it points to address in RAM but section is placed into
FLASH.

This is not a problem in current design because absolute address
attribute is not used for the section, but this will be a problem in
future improvements of linker script, so this issue is fixed as a
preparation step.

To fix the issue sections are grouped together (all FLASH sections are
defined before first RAM section) which ensures that location counter is
always going up and is correctly updated.

Fix is not needed for level 3 linker script because it already has
sections in correct order.

Signed-off-by: Bohdan Hunko <Bohdan.Hunko@infineon.com>
Change-Id: I40172f20b83fa4447f7e9ba4cc81dc21f070c0e2