Platform: Fix sections alignments
Following is simplified notation for GCC linker script section:

    section [address] [(type)] : [ALIGN(section_align)]
    {
        output-section-command
        ...
    } [>region] [AT>lma_region]

If [address] is present then it will be used as address for the output
section. If [address] is not present then section will be placed into
[>region] and aligned to the strictest alignment of any input section
contained within the output section.
This means that if section contains data which has
__attribute__(align( align_size )) attribute it will effect alignment of
the section. So in following example:

    section1 : ALIGN(4)
    {
        .test_section_one
        .test_section_two
        .section_with_some_variable /* some_variable is 1KBs aligned */
        .test_section_three
    } > FLASH

section1 will be 1KBs aligned even though ALIGN(4) is specified.
This results is wasted memory. More over despite the fact that section1
will be 1KB aligned *fill* will still be added before some_variable
to make it 1KB aligned. So this is even more wasteful.

For more information on this topic refer to
"3.6.1 Output Section Description" and "3.6.3 Output Section Address"
paragraphs of ARM GCC Linker document.

This patch fixes the problem by setting [address] instead of
[ALIGN(section_align)]. [address] is set to value of location counter
aligned to needed alignment. This results in section being placed at
exactly specified address without any unnecessary paddings.

[address] attribute is not used for the first section (TFM_VECTORS)
because at that place value of location counter is 0.

Signed-off-by: Bohdan Hunko <Bohdan.Hunko@infineon.com>
Change-Id: I4abe5d0ff43414575ab7cb2fc7fd87c9cc909252