SPM: Do not use msg.in_size to record data read/skipped
The SPM now uses the msg.in_size within the connection_t struct to
record the remaining input data size.

There will be a problem in the SFN backend because the "msg" in the
connect_t struct is directly passed to RoT Services.
Then RoT Services might see a change of the msg->in_size after a call to
psa_read or psa_skip.
This is not the expected behaviour. The in_size should not be changed
within a call to the RoT Services.
This is an known issue in TF-M v1.8.0, see
https://developer.trustedfirmware.org/T1030 for details.

This patch fixes this issue.
It expands the invec and outvec struct in connect_t struct and keep the
"base" field only because the "len" in the psa_invec/outvec has
different meanings in the connect_t.
To avoid confusions, new dedicated fields are added to record data size
accessed and written.

Note that the IPC backend does not have this issue because there is
always a local copy of the msg_t struct.

Change-Id: I82b4a7821e963697c8eef7de15e91aae2c662966
Signed-off-by: Kevin Peng <kevin.peng@arm.com>