RSS: Add Secure Data Storage (SDS) driver
Add a reduced version of SDS driver to RSS to share
data with AP through shared memory. The point is to
have an offline data sharing solution which does
not rely on the MHU communication. The driver API
is available via platform specific SVC call.

The SDS region is platform specfic and defined at
build time on both sides: RSS and AP.

Reduced features of SDS:
 - Adding finalized data structures is supported only.
 - Read or partial read from data structure: Not supported
 - Write or partial write to data structure: Not supported

Signed-off-by: Tamas Ban <tamas.ban@arm.com>
Signed-off-by: David Vincze <david.vincze@arm.com>
Change-Id: I895860ee9fe0114ced66fe5c2247d02ff46b7b00