Build: GCC: Add NOLOAD attribute for BSS sections
The output sections of name with "BSS" are expected to put only
(*.bss*) and (COMMON) input sections so that the type will be SHT_NOBITS
and the sections do not occupy space in the output file because there is
no need to put zeros in output files.

Whilst there are two exceptions, they are the
*(TFM_*_APP-ROT_ATTR_ZI) and *(TFM_*_PSA-ROT_ATTR_ZI) input sections.
These might be done by mistake and make the corresponding output
sections become PROGBITS type which takes space.

Note there are no usages of these sections thus the issue does not
emerge.

This patch fixes this issue by adding "NOLOAD" for BSS sections.
This explicitly make the output sections "NOBITS".

Change-Id: I1ecb25763436989f407498d3b3e3e718c51710c5
Signed-off-by: Kevin Peng <kevin.peng@arm.com>