Docs: Minor rewording of dual core design doc
Primarily to replace the use of the word "should" when describing
mandatory aspects. Requirements documents generally use "should" to
identify requirements that are desirable but not mandatory, using
"shall", "must", or "will" for mandatory requirements, so its useful to
follow this same practice here.

I also reworded a few sentences in minor ways.

Signed-off-by: Chris Brand <chris.brand@cypress.com>
Change-Id: I57b8bcadf01ffd638c9ac2c26db7ac65cc704c37