Build: move signature key len into signature type
The current approach in MCUBoot assumes the signature type and the
key len are separate values, but this only applies to RSA. For
other signature types (such as ECDSA) the key len is not applicable.
This change merges the key length and the signature type for RSA
thus removing the need for extra conditions in the build
configuration when the signature is specified in a different way.

Signed-off-by: Roland Mikhel <roland.mikhel@arm.com>
Change-Id: I5e2bba814695ec0d366b202b9c40fd7daa9484ac