Platform: AN552: Use provisioning bundles
If "-DPLATFORM_DEFAULT_PROVISIONING=OFF" then use provisioning
bundles. A new binary is built that contains the provisioning values
and code. All three of the previous provisionings are moved to the
united provisioning bundle.

The approach is based on RSS's provisioning, but the OTP layout,
provisioning values have changed and it uses flash based OTP.

Change-Id: I68710a3c037f0a17245849ea6b01f4b3ba16945e
Signed-off-by: Bence Balogh <bence.balogh@arm.com>