Build: Minor fix
In case of TFM_MULTI_CORE_TOPOLOGY, the tfm_ipc_config_default.cmake is
always included. So the CMAKE variable CONFIG_TFM_SPM_BACKEND_SFN which
is only set in the tfm_sfn_config_default.cmake will not be created
if TFM_MULTI_CORE_TOPOLOGY is ON. So the invalid check of the setting
(TFM_PARTITION_NS_AGENT_MAILBOX AND CONFIG_TFM_SPM_BACKEND_SFN) will
always pass even when CONFIG_TFM_SPM_BACKEND is set to "SFN" with
multi-core platforms. This patch is to fix this issue.

Signed-off-by: Sherry Zhang <sherry.zhang2@arm.com>
Change-Id: I8246796e3ddb4d2b8e246d37967cc64c223793cc