platform: nordic_nrf: Use aliased default IRQ handlers
Currently we are generating dozens of default IRQ Handlers that each
just spin. In sum, 256 bytes of IRQ handlers that do nothing.

In this patch we save these bytes by using the alias attribute.

Now there is only one default IRQ handler placed in flash and all the
default IRQ handlers alias to it.

I tried to place the duplicated default_tfm_IRQHandler in the common
startup.c, but it did not work. I am not certain, but I think this is
due to a problem with the way we link in TF-M.

Change-Id: Ifd3a67ac2276fa8d7ceec29482f8cec02b2cbd54
Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>