platform: nordic_nrf: Don't hang in IRQs by default
The current default IRQ handler will hang forever. It is bad practice
to hang forever in production.

Instead we now print the exception information and then panic. Which
will either reboot or hang forever depending on the configuration.

It is bad practice to for the default IRQ handler to be triggered at
all and it should indicate a configuration error.

Change-Id: Ifd3a67ac2276fa8d7ceec29482f8cec02b3cac54
Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>