Platform: AN552: Create unique provisioning bundle
If TFM_DUMMY_PROVISIONING is OFF then calculate the
hash of the new private keys' public pair and use that for
provisioning.
The other keys can also be generated if needed or can be
set in the config files.

The provisioning bundle build is supported by armclang, gcc
and iar as well.

Change-Id: I32af4a34720cfa0aeda1220499aa9404a0056cad
Signed-off-by: Dávid Házi <david.hazi@arm.com>
Signed-off-by: Bence Balogh <bence.balogh@arm.com>