Docs: Simplify Mailbox NS Agent work behavior
The initial design covers the case that a service access reply might
be returned already when NS Agent gets scheduled back, for example:

  1. After agent_psa_call, both agent and target service are runnable.
  2. Schedule, and the service gets run.
  3. Service replies, then the agent gets its return value just while
     returning from agent_psa_call.
  4. Then many checkings need to be added to decide if the ACK signal
      should be set or not, which complicates the logic.

Here forcing the replies for agent API to be fetched by a `psa_get` can
simplify the logic. Updated the pseudo-code in the design document for
this, and the status code for `in-progress` is not needed anymore.

Signed-off-by: Ken Liu <Ken.Liu@arm.com>
Change-Id: Icf69328445aac5866f611364b1cf3b2354405fd8