RSS: Add NI-Tower discovery driver
Arm CoreLink NI-Tower Network‑on‑Chip Interconnect is a highly
configurable system-level non‑coherent interconnect that is compliant
with AMBA protocols and is optimized to the Power, Performance, and Area
(PPA) requirements of a SoC design.

The NI-Tower interconnect has a configurable topology to support various
SoC layout requirements. The Discovery feature of NI-Tower allows
software to determine the structure of the NI-Tower domains, components,
and subfeatures on system boots.

As part of initial driver support for the NI-Tower, add support to
discover the address offset of a subfeature node with given subfeature
node type along with its parent component node id and node type.

Signed-off-by: Arnold Gabriel Benedict <arnoldgabriel.benedict@arm.com>
Change-Id: I07e145d1ecdcd4e0d8151b2e300695e7cbbc33fa