Platform: RSS: Replace image number macro with condition
The image number macro neatly makes the values in the table conditional,
but it generates a lot of spurious compiler warnings due to accessing
members that do not exist with lower image numbers, which can make real
warnings hard to spot in the build log. Replace it with plain
preprocessor conditions that achieve the same goal without warnings.

Signed-off-by: Jamie Fox <jamie.fox@arm.com>
Change-Id: I8d7b6fb2356c8f24cfa15fa0ed2b89f5cb73e88c