Platform: Use NS flag only for ns_agent_tz in isolation HAL
The NS flag in the Armv8-M isolation HAL is used to determine if the NS
MPU should be checked. This check only makes sense for a call from the
TrustZone NSPE, so only set this flag for ns_agent_tz and not
ns_agent_mailbox. This change makes it possible to use the HAL on
platforms with both agents.

Signed-off-by: Jamie Fox <jamie.fox@arm.com>
Change-Id: I97be3e964008f333993df8e52f46e3420aacb9f2