SPM: Distinguish clients in psa_connect/close in dual-core
In psa_close, it checks the current/input client_id with the
value that is stored in the connection handle. So the
client ID information should be got from the nonsecure
core in psa_close. agent_psa_close is added for dual-core
with client_id parameter.

Pass the input client_id from the request via the mailbox
into SPM in psa_connection and psa_close. The implementation
of tfm_spm_client_psa_connect and tfm_spm_client_psa_close
is adjusted to make it unified in the agent case and the other
case.

Signed-off-by: Sherry Zhang <sherry.zhang2@arm.com>
Change-Id: Iacc9a48b0f0351a2d93c5c1499ae915bd8669e45