SPM: Rename Cross Call to Thread Function Call
The name cross-call describes the implementation of SPM in
isolation level 1. A thread requests the services from other secure
partition threads via PSA API. It is required to switch to the SPM
stack to avoid using up the Secure Partition's stack. 'cross' here
describes 'cross-stack'.

However, the SVCall in isolation level 2/3 also requires crossing
stack and even crossing from thread mode to handler mode. So, the
cross-call is not an accurate description of the operation for
the situation in isolation level 1.

Using 'Thread Function Call' to describe the operation is better
because it happens in the Thread Mode and it is a direct function
call. These related contents are updated as well.

Signed-off-by: Jianliang Shen <jianliang.shen@arm.com>
Change-Id: I1c6b32e3166cbca7d3ae1a4c6ae8485dce935fed