platform: nordic_nrf: Fix assert in initialization
This fixes an erroneous attempt to configure the FPU
peripheral in NRF91 series. The FPU is not configurable
in NRF91 series and it is always non-secure. This caused
a failed assertion in the initialzation since the FPU
present field is not set for the SPU configuration.

This change configures the FPU only in NRF53 series devices
which have a configurable FPU.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Change-Id: Ia00e0e620ab51850e6ee8b382e5d03a245b6db0c