config: Add invalid config for MCUBOOT_HW_KEY
If the MCUBOOT_HW_KEY flag is OFF then the BL ROTPKs
have to be set compile time and cannot be provisiond runtime.
Currently the dummy BL2 S/NS signing keys are hardcoded in the
keys.c so the dummy private keys have to be used for signing
the images so we cannot generate and use unique private keys.

Change-Id: I6da21370f1b27b3bd023f6b82bbd3f871cbd5a7d
Signed-off-by: Bence Balogh <bence.balogh@arm.com>