SPM: Save CONTROL register before switching to Thread mode
When preparing to switch to SPM implementation in Thread mode, the
CONTROL register is changed to make Thread mode privileged.

This patch adds saving of the current CONTROL register value and
restoring of it after the return from Thread mode. This is necessary
because when the scheduler runs after this, it only calls
tfm_hal_activate_boundary() to change the CONTROL register if the next
partition to run is in a different boundary to the current one. So, the
CONTROL register needs to contain the correct value for the current
partition in case it is reused for the next partition.

Signed-off-by: Jamie Fox <jamie.fox@arm.com>
Change-Id: If6f47ac0152e7bc35af641f2fdd1573c3dbb19d1