Platform: RSS: mhu: add signal and wait api
Add signal_and_wait_for_clear and wait_for_signal_and_clear public
functions. The first function allows the sender to send a message and
wait for the receiver end to acknowledge and the second function allows
the receiver to wait for the sender to send a message and clear once the
message is received. Both functions will use the last available channel
to do its functionality.

MHUv2 wrapper already have implementation of signal_and_wait_for_clear
function. This patch exposes them as public functions with necessary
changes to align with the api.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: I9a855684f63fc01e62fe3bacab8ca4da0ad56266