SPM: Manage SPM boundary during psa api execution
When triggering SVC handler to execute psa api execution,
this means that it has entered SPM domain, so it's necessary
to judge whether it needs to switch boundary between SPM and
current partition.
In the same way, the same judgement is needed when psa api
execution ends before returning to previous partition.

Signed-off-by: Chendi Sun <chendi.sun@arm.com>
Change-Id: I28e9fae3b2105022695ae7df14d063c6940e0537