RSE: Fremont: Add sideband MHU devices
Sideband MHUs are used for cross chip RSE to RSE communications
before CMN has been setup in the multichip scenario.

As the sideband MHUs are MHUv2 but all other MHUs in the system are
MHUv3 the common cmake files which add either MHUv2 or MHUv3 to all
firmware cannot be used. Instead the common cmake MHUs files are
disabled by setting PLAT_MHU_VERSION to 0 (a value unlikely to ever be
used) and instead the sub-platform make files add MHUv2 to BL1 and MHUv3
to all other firmware.

Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: Ib9eba479d01d48f685abc70122376e2ee6860f48