SPM: Introduce uniform Client ID mapping
This commit adds a mapping function to translate Client IDs provided by
Non-Secure clients to Client IDs that are unique in SPM. To achieve this
a Client ID base and a limit value is added to the NS agent partitions'
manifests. The mapping function ensures that the mapped Client ID value
is in the range defined for that NS agent partition. SPM also makes sure
in boot time that the Client ID regions defined in the manifests are not
overlapping.

The Client ID provided by the NS client '-1 * n' must have the following
properties:
 - n > 0
 - n <= (-1 * client_id_base) - (-1 * client_id_limit) + 1

Signed-off-by: Sherry Zhang <sherry.zhang2@arm.com>
Signed-off-by: Mate Toth-Pal <mate.toth-pal@arm.com>
Change-Id: I5a9ad1db3458372eb5706457be641a5e50178728