RSE: Fremont: Configure multichip topology
Configure topology file used for generating the OTP routing map used for
VHUK key sharing.

The topology file contains a description of which RSE nodes exist in the
graph and how they are connected to each other represented using "trival
graph format". The connections are then labeled with the index of MHU
that should be used to send and receive from that connection in the
platform specific MHU_RSE_TO_RSE_RECEIVER_DEVS and
MHU_RSE_TO_RSE_SENDER_DEVS arrays.

Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: Ic004ac11364cea287f04a58a3d7e5a4cdd18fbd2