SPM: Print the TF-M version as early as possible
For debug purposes it is good practice for the boot banner to be the
first message that is logged.

The earliest point that we can start logging is after the platform has
initialized it's logging backend, like a UART driver.

We therefore move the boot banner from after tfm_core_init to just
after tfm_hal_platform_init.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>
Change-Id: I771f8fa0867382ee6a688322b91943a247d1fd72