Dual-core: Remove limitation of single mailbox slot in NS bare metal env
Though it wastes some resources to enable multiple mailbox slots in NS
bare metal environment, it makes it possible to use a TF-M secure image
in different NS use cases. It should be NS developers who decide how
many slots to be used.

So only throw a message to remind developers that number of mailbox
slots is set to more than 1 in NS bare metal environment.

Signed-off-by: Xinyu Zhang <xinyu.zhang@arm.com>
Change-Id: I7ba8c2c207547d916a6c07aa97ffe0e1c55b6cfb