interface: Add workaround for mbedtls using PSA
Add workaround for mbedtls using PSA.
The MbedTLS library is not compatible with the PSA headers provided
by TF-M. Add prototype of the functions not supported by TF-M to prevent
undefined reference warning. This will still fail at link-time if the
application tries to use the functions.

This commit is temporary until mbedtls and TF-M projects address this
issue.
https://github.com/Mbed-TLS/mbedtls/issues/8602
Fix is in mbed TLS commit:
8f1c36df00f6cdca0ba0bb04a497057c37248017
Could not cherry-picked due to it beeing a merge commit

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>
Change-Id: I66fcb9d971929f3c21e93c8c3f3c0c666c6563c7
Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>