Build: Choose IPC model for platforms with TZ & mailbox
Changes the build system to choose IPC model by default for platforms
that use both TrustZone and the mailbox agent (hybrid platforms).

Although support for the mailbox agent was added to SFN model in
415d713, the SFN model does not currently allow execution to return to
the NSPE when the mailbox agent calls psa_wait() to wait for the mailbox
interrupt, so IPC is a better choice in this case.

Signed-off-by: Jamie Fox <jamie.fox@arm.com>
Change-Id: Idd265ce60ca15dcfd72ca65435c9a9ea74a20ad5