SPM: Overwrite pool data on free in debug builds
Changes tfm_pool_free() to fill the freed pool chunk with 0xFF, only in
debug builds. This makes it more likely for a use-after-free bug (such
as the one previously fixed in e85d75bd0) to cause an error in debug
builds and so be caught before making it into released code.

Signed-off-by: Jamie Fox <jamie.fox@arm.com>
Change-Id: I915fd4e458478e3a3e04c9ea685f4ebca7d261ad