platform: nordic_nrf: Remove unused driver functions
Remove unused driver functions.

We are currently implementing several functions that TF-M is not
using.

This is bad practice as they are untested and may therefore be
unreliable if TF-M were to start to use them.

They also bloat the size of the binary and have a code maintenance
cost.

It would be better to implement the functions when they become used.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>
Change-Id: Icd7df7caa38ea890742b4b70118d642b196c4d71