RSS: Fremont: Fix RSS load non-secure image guards
The guard statments around RSS_FIRMWARE_NON_SECURE_ID in the BL2 load
vector tables are wrong. Currently they only add the vector function
when RSS_LOAD_NS_IMAGE is undefined and the funciton is not called. This
causes a crash when RSS_LOAD_NS_IMAGE is defined. The preload function
is only called when RSS_LOAD_NS_IMAGE is defined however the functions
are removed when RSS_LOAD_NS_IMAGE is defined so it tries to call a null
pointer as a function.

So invert the guards on RSS_FIRMWARE_NON_SECURE_ID vector table so the
function is only present when RSS_LOAD_NS_IMAGE is defined.

Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: I4d340fbb4d38a77876843d59784ab8198212eabc