RSE: host_drivers: Fix NI-Tower APU re-configuration issues
When reconfiguring an APU region, previously set permissions or APU ids
were not cleared out, leading to wrong configuration of permissions or
APU ids. Resolve this issue by clearing out the permissions and APU ids
before programming the new values.

Also, if a region is locked, re-programming a region should not be
allowed to avoid any exceptions raised by the NI-Tower. Add a check to
disallow re-configuration when a region is locked.

Signed-off-by: Arnold Gabriel Benedict <arnoldgabriel.benedict@arm.com>
Change-Id: I9c1f9aa18e20b79f8acaa4d59fd264647ecc6e2e