RSE: host_drivers: Check region overlap for NI-tower subfeature configs
When configuring PSAM or APU regions, the memory region entry could
overlap with another region, which is already enabled in the PSAM or
APU. Check if the base address and end address of the new entry
overlaps with existing/enabled entry in the node. In case of APU, two
foreground regions or two background region should'nt overlap. Whereas,
a foreground region can overlap two background region. The APU
priorities the foreground access permission in such case.

Signed-off-by: Arnold Gabriel Benedict <arnoldgabriel.benedict@arm.com>
Change-Id: I6eb69486b75c10cdf80f54740889cc64239f7210