Build: Remove input of standard libraries in linker scripts
Removes the GROUP() command with standard libraries in the GCC linker
scripts in favour of specifying -specs=nano.specs and -specs=nosys.specs
in the toolchain file. These two specs together cause the binary to be
linked with "-lgcc -lc_nano -lnosys" which is exactly what we want. Also
specifying this in the linker script is mostly duplication, but actually
somewhat detrimental in some cases as the GROUP() command included libc,
whereas we prefer to only use libc_nano.

Also removed -nostdlib from the compile options as it is a linker option
and so had no effect there. It cannot be used as a link option with TF-M
currently since we do rely on a few things from the standard libraries.

Signed-off-by: Jamie Fox <jamie.fox@arm.com>
Change-Id: I190cad7f112df5de75417776c42af7bb9a9d9c35