RSE: Fremont: Add rse_scp_axis PSAM config
Add PSAM address map to map access from RSE and SCP targeting LCP
address space. All access from RSE and SCP that target to LCP is routed
via a NIC400 outside of System Control NI-Tower and sent to the
'rse_scp_axis' on the NI-Tower.

This is necessitated because address filtering on transactions based on
the LCP memory map is required in the NI-Tower, but the target
'lcp_axim' interface APU can only do address filtering for upto 8 LCPs.
Hence, for every transaction targeting the LCP, the upper address bits
are propagated via AXI USER bits rather than on the AXI address signals
and the address filtering is applied based on the lower address bits.
With this, NI-Tower has to handle only on a 'single' LCP in the system,
regardless of the number of LCPs in the system or regardless of the
chip address offset.

Signed-off-by: Arnold Gabriel Benedict <arnoldgabriel.benedict@arm.com>
Change-Id: Ib203ea14e5f96e6810dd5dae29f00aec808a533c