RSE: Fremont: Add app_axis PSAM config
Add PSAM region mappings for the app_axis interface of the System
Control NI-Tower instance. The physical address coming out of AP will
be mapped to the target interface based on these PSAM region mappings.
Since this interface is under the SYSTOP power domain and not under
the AON power domain, a new function is introduced to configure the
PSAMs under the SYSTOP domain and app_axis will be programmed as part
of it.

Note: The interfaces under SYSTOP domain can be accessed only after
SYSTOP is initialized. Accessing these node before SYSTOP init will
lead to error.

Signed-off-by: Arnold Gabriel Benedict <arnoldgabriel.benedict@arm.com>
Change-Id: I29dfd8a14f87d5d5ddf51850c419a5e532712536