RSE: Fremont: Configure SCP-RSE MHUs
Add initilization function to configure the RSE to SCP MHU and the SCP
to RSE MHU. Also add interrupt handler function for receiving messages
from SCP to RSE MHU and set the handler in the IRQ vector table.
This function will call one of the handler functions depending on the
MHU channel used. Currently no channels are configured.

Signed-off-by: Shriram K <shriram.k@arm.com>
Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: I4738bc7c4c924880b43e71b49310ed9c07668c0c