RSE: Fremont: Wait for SCP signal before LCP load
Before MCUBoot can load LCP firmware, the SYSTOP power domain needs to
be in ON state. The SYSTOP power domain is switched on by SCP, so RSE
has to wait for SCP to signal when the SYSTOP init is complete, before
MCUBoot can load the LCP firmware.

When SCP has initialized SYSTOP, it sends an MHU message to RSE via the
SCP to RSE MHU. Hence, configure RSE to wait for this message before
letting MCUBoot try to load the LCP firmware image.

In order to monitor the status of the host systems such as SCP and AP,
as well as to orchestrate the comms between RSE and host systems,
introduce host system as abstraction layer. Configure the SCP to RSE MHU
interrupt to notify host system module when the SYSTOP ready signal is
received so host system can be queried about whether SYSTOP is ready.

Signed-off-by: Shriram K <shriram.k@arm.com>
Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: I4cb2f008adc08d7531b614a08c3a97e8d0b72a2d