RSE: Fremont: Configure BL2 for multichip
Multichip variants of RD-Fremont need each RSE to load firmware to its
local chip.

SCP and MCP address spaces are local to the chip, therefore no change is
needed to access the local SCP and MCP.

For AP, each chip is allocated a 64GiB local address space, and each
chip's address space is present at a 64GiB offset from the start of the
previous chip's address space.

In order to access the host AP address space of a specific chip, add a
offset value of 64GiB * chip_id to the existing host memory maps. For
example, when RSE in the Chip 2 needs to access its host address space,
it adds an offset of 128GiB when setting up the physical address of the
ATU. The chip_id is determined by reading the config value from the OTP
in the host system model. The correct chip_id is expected to be
provisioned to the OTP during the Device Manufacturing stage.

Also, loading of AP BL1 image is required only for the Chip 0 as in a
multichip scenario, only the primary core of the Chip 0 is booted, the
secondary cores are brought up by the OSPM.

Signed-off-by: Shriram K <shriram.k@arm.com>
Signed-off-by: Pranav Madhu <pranav.madhu@arm.com>
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Signed-off-by: Arnold Gabriel Benedict <arnoldgabriel.benedict@arm.com>
Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: Id71be8f311633dbd3bb8248a4ec64b14f0545301