RSE: Fremont: Support for RD-Fremont-Cfg2 variant
Add support for RD-Fremont-Cfg2 platform which is a quad-chip variant of
the RD-Fremont platform. The FVP implementation of RD-Fremont-Cfg2 has
reduced AP core count per chip and hence LCP core count is also reduced.
Also, RSE_AMOUNT is set to 4, so that RSE firmware is configured to
enable the multi RSE specific features such as RSE to RSE sideband
communication and multi RSE vHUK generation and exchange.

Signed-off-by: Shriram K <shriram.k@arm.com>
Signed-off-by: Pranav Madhu <pranav.madhu@arm.com>
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Signed-off-by: Arnold Gabriel Benedict <arnoldgabriel.benedict@arm.com>
Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: Idc6fb0bc37e8f07b2aed93295fdf1dccf1deb763