RSE: Fremont: Release MCP and SCP CPU from halt
SCP and MCP are in a halted state and RSE is responsible for releasing
it from halt after the image loading to respective ITCM is completed.
This is done by configuring the MSCP_CPU_INIT_CTRL in the SCP RAS and
INIT Control registers address space.

Use MSCP drivers to unhalt SCP and MCP during post init stages of
respective firmware loading.

Signed-off-by: Arnold Gabriel Benedict <arnoldgabriel.benedict@arm.com>
Signed-off-by: Shriram K <shriram.k@arm.com>
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: I94ba77ec4ecfc95f71626db55a691a7b08008168