RSE: Fremont: Configure System Control NI-Tower under SYSTOP
NI-Tower has programmable address maps (PSAM) and Access Protection
Units (APU) which can be configured during boot time which it uses to
map system addresses to interconnect targets for routing purposes as
well as to specify access permissions to configured regions.

Several nodes in the System Control NI-Tower instances are under SYSTOP
power domain rather than on the AON domain. So configure PSAMs and APUs
under SYSTOP power domain after RSE completes loading of SCP. The
support for RSE to wait for the SCP to complete the SYSTOP init will be
introduced in the subsequent patches.

Signed-off-by: Arnold Gabriel Benedict <arnoldgabriel.benedict@arm.com>
Change-Id: If078673a1a8d097d40357ef41f8534537ee93bdc