Mailbox: Return status from agent psa close calls
Changes the agent_psa_close() and tfm_rpc_psa_close() functions to
return the error status returned from the SPM. Even though the
psa_close() API does not return an error value, the SPM functions can
fail due to a programmer error (invalid handle or partition not idle).
As the caller is NS, the programmer error is returned by SPM rather than
blocking execution.

This patch propagates the returned error to the mailbox implementation,
which makes it possible to send an immediate reply to the mailbox
caller, as no asynchronous reply will be received for these failed
calls.

Signed-off-by: Jamie Fox <jamie.fox@arm.com>
Change-Id: I73689123000a9e357effe0b4522223e9f4338c76